refactor: create UI slice
