+++ api-editor/gui/src/common/MenuBar.tsx
@@ -28,13 +28,14 @@ import {
 import React, { useRef, useState } from 'react';
 import { FaChevronDown } from 'react-icons/fa';
 import { useAppDispatch, useAppSelector } from '../app/hooks';
-import { resetAnnotations, toggleAnnotationImportDialog } from '../features/annotations/annotationSlice';
+import { resetAnnotations } from '../features/annotations/annotationSlice';
 import AnnotatedPythonPackageBuilder from '../features/annotatedPackageData/model/AnnotatedPythonPackageBuilder';
 import PythonPackage from '../features/packageData/model/PythonPackage';
 import { HeatMapMode, setHeatMapMode, togglePackageDataImportDialog } from '../features/packageData/packageDataSlice';
 import { Setter } from './util/types';
 import { toggleUsageImportDialog } from '../features/usages/usageSlice';
 import { FilterHelpButton } from './FilterHelpButton';
+import {toggleAnnotationImportDialog} from "../features/ui/uiSlice";
 
 interface MenuBarProps {
     pythonPackage: PythonPackage;
