+++ api-editor/gui/src/features/annotations/annotationSlice.ts
@@ -15,7 +15,6 @@ export interface AnnotationsState {
     constants: {
         [target: string]: ConstantAnnotation;
     };
-    currentUserAction: UserAction;
     enums: {
         [target: string]: EnumAnnotation;
     };
@@ -40,7 +39,6 @@ export interface AnnotationsState {
     removes: {
         [target: string]: RemoveAnnotation;
     };
-    showImportDialog: boolean;
 }
 
 export interface AttributeAnnotation {
@@ -258,64 +256,6 @@ export interface RemoveAnnotation {
     readonly target: string;
 }
 
-type UserAction =
-    | typeof NoUserAction
-    | AttributeUserAction
-    | BoundaryUserAction
-    | CalledAfterUserAction
-    | ConstantUserAction
-    | GroupUserAction
-    | EnumUserAction
-    | RenameUserAction
-    | OptionalUserAction;
-
-const NoUserAction = {
-    type: 'none',
-    target: '',
-};
-
-interface AttributeUserAction {
-    readonly type: 'attribute';
-    readonly target: string;
-}
-
-interface BoundaryUserAction {
-    readonly type: 'boundary';
-    readonly target: string;
-}
-
-interface CalledAfterUserAction {
-    readonly type: 'calledAfter';
-    readonly target: string;
-    readonly calledAfterName: string;
-}
-
-interface ConstantUserAction {
-    readonly type: 'constant';
-    readonly target: string;
-}
-
-interface EnumUserAction {
-    readonly type: 'enum';
-    readonly target: string;
-}
-
-export interface GroupUserAction {
-    readonly type: 'group';
-    readonly target: string;
-    readonly groupName: string;
-}
-
-interface OptionalUserAction {
-    readonly type: 'optional';
-    readonly target: string;
-}
-
-interface RenameUserAction {
-    readonly type: 'rename';
-    readonly target: string;
-}
-
 // Initial state -------------------------------------------------------------------------------------------------------
 
 export const initialState: AnnotationsState = {
@@ -323,7 +263,6 @@ export const initialState: AnnotationsState = {
     boundaries: {},
     calledAfters: {},
     constants: {},
-    currentUserAction: NoUserAction,
     enums: {},
     groups: {},
     moves: {},
@@ -331,7 +270,6 @@ export const initialState: AnnotationsState = {
     pures: {},
     renamings: {},
     requireds: {},
-    showImportDialog: false,
     removes: {},
 };
 
@@ -478,72 +416,7 @@ const annotationsSlice = createSlice({
         },
         removeRemove(state, action: PayloadAction<string>) {
             delete state.removes[action.payload];
-        },
-        showAttributeAnnotationForm(state, action: PayloadAction<string>) {
-            state.currentUserAction = {
-                type: 'attribute',
-                target: action.payload,
-            };
-        },
-        showBoundaryAnnotationForm(state, action: PayloadAction<string>) {
-            state.currentUserAction = {
-                type: 'boundary',
-                target: action.payload,
-            };
-        },
-        showCalledAfterAnnotationForm(state, action: PayloadAction<CalledAfterTarget>) {
-            state.currentUserAction = {
-                type: 'calledAfter',
-                target: action.payload.target,
-                calledAfterName: action.payload.calledAfterName,
-            };
-        },
-        showConstantAnnotationForm(state, action: PayloadAction<string>) {
-            state.currentUserAction = {
-                type: 'constant',
-                target: action.payload,
-            };
-        },
-        showGroupAnnotationForm(state, action: PayloadAction<GroupTarget>) {
-            state.currentUserAction = {
-                type: 'group',
-                target: action.payload.target,
-                groupName: action.payload.groupName,
-            };
-        },
-        showEnumAnnotationForm(state, action: PayloadAction<string>) {
-            state.currentUserAction = {
-                type: 'enum',
-                target: action.payload,
-            };
-        },
-        showMoveAnnotationForm(state, action: PayloadAction<string>) {
-            state.currentUserAction = {
-                type: 'move',
-                target: action.payload,
-            };
-        },
-        showOptionalAnnotationForm(state, action: PayloadAction<string>) {
-            state.currentUserAction = {
-                type: 'optional',
-                target: action.payload,
-            };
-        },
-        showRenameAnnotationForm(state, action: PayloadAction<string>) {
-            state.currentUserAction = {
-                type: 'rename',
-                target: action.payload,
-            };
-        },
-        hideAnnotationForms(state) {
-            state.currentUserAction = NoUserAction;
-        },
-        toggleImportDialog(state) {
-            state.showImportDialog = !state.showImportDialog;
-        },
-        hideImportDialog(state) {
-            state.showImportDialog = false;
-        },
+        }
     },
     extraReducers(builder) {
         builder.addCase(initializeAnnotations.fulfilled, (state, action) => action.payload);
@@ -579,20 +452,6 @@ export const {
     removeRequired,
     addRemove,
     removeRemove,
-
-    showAttributeAnnotationForm,
-    showBoundaryAnnotationForm,
-    showCalledAfterAnnotationForm,
-    showConstantAnnotationForm,
-    showEnumAnnotationForm,
-    showGroupAnnotationForm,
-    showMoveAnnotationForm,
-    showOptionalAnnotationForm,
-    showRenameAnnotationForm,
-    hideAnnotationForms,
-
-    toggleImportDialog: toggleAnnotationImportDialog,
-    hideImportDialog: hideAnnotationImportDialog,
 } = actions;
 export const annotationsReducer = reducer;
 
@@ -613,7 +472,6 @@ export const selectConstant =
     (target: string) =>
     (state: RootState): ConstantAnnotation | undefined =>
         selectAnnotations(state).constants[target];
-export const selectCurrentUserAction = (state: RootState): UserAction => selectAnnotations(state).currentUserAction;
 export const selectEnum =
     (target: string) =>
     (state: RootState): EnumAnnotation | undefined =>
@@ -642,8 +500,6 @@ export const selectRequired =
     (target: string) =>
     (state: RootState): RequiredAnnotation | undefined =>
         selectAnnotations(state).requireds[target];
-export const selectShowAnnotationImportDialog = (state: RootState): boolean =>
-    selectAnnotations(state).showImportDialog;
 export const selectRemove =
     (target: string) =>
     (state: RootState): RemoveAnnotation | undefined =>
