+++ api-editor/gui/src/features/annotations/forms/BoundaryForm.tsx
@@ -20,9 +20,10 @@ import { useForm } from 'react-hook-form';
 import { useAppDispatch, useAppSelector } from '../../../app/hooks';
 import { numberPattern } from '../../../common/validation';
 import PythonDeclaration from '../../packageData/model/PythonDeclaration';
-import { ComparisonOperator, hideAnnotationForms, Interval, selectBoundary, upsertBoundary } from '../annotationSlice';
+import { ComparisonOperator, Interval, selectBoundary, upsertBoundary } from '../annotationSlice';
 import { AnnotationForm } from './AnnotationForm';
 import { Optional } from '../../../common/util/types';
+import {hideAnnotationForm} from "../../ui/uiSlice";
 
 interface BoundaryFormProps {
     readonly target: PythonDeclaration;
@@ -93,11 +94,11 @@ export const BoundaryForm: React.FC<BoundaryFormProps> = function ({ target }) {
                 },
             }),
         );
-        dispatch(hideAnnotationForms());
+        dispatch(hideAnnotationForm());
     };
 
     const onCancel = () => {
-        dispatch(hideAnnotationForms());
+        dispatch(hideAnnotationForm());
     };
 
     // Validation ------------------------------------------------------------------------------------------------------
