+++ api-editor/gui/src/features/annotations/forms/EnumForm.tsx
@@ -14,8 +14,9 @@ import { FaPlus, FaTrash } from 'react-icons/fa';
 import { useAppDispatch, useAppSelector } from '../../../app/hooks';
 import { pythonIdentifierPattern } from '../../../common/validation';
 import PythonDeclaration from '../../packageData/model/PythonDeclaration';
-import { hideAnnotationForms, selectEnum, upsertEnum } from '../annotationSlice';
+import { selectEnum, upsertEnum } from '../annotationSlice';
 import { AnnotationForm } from './AnnotationForm';
+import {hideAnnotationForm} from "../../ui/uiSlice";
 
 interface EnumFormProps {
     target: PythonDeclaration;
@@ -101,11 +102,11 @@ export const EnumForm: React.FC<EnumFormProps> = function ({ target }) {
                 ...data,
             }),
         );
-        dispatch(hideAnnotationForms());
+        dispatch(hideAnnotationForm());
     };
 
     const onCancel = () => {
-        dispatch(hideAnnotationForms());
+        dispatch(hideAnnotationForm());
     };
 
     // Rendering -------------------------------------------------------------------------------------------------------
