+++ api-editor/gui/src/features/annotations/forms/MoveForm.tsx
@@ -4,8 +4,9 @@ import { useForm } from 'react-hook-form';
 import { useAppDispatch, useAppSelector } from '../../../app/hooks';
 import { moduleNamePattern } from '../../../common/validation';
 import PythonDeclaration from '../../packageData/model/PythonDeclaration';
-import { hideAnnotationForms, selectMove, upsertMove } from '../annotationSlice';
+import { selectMove, upsertMove } from '../annotationSlice';
 import { AnnotationForm } from './AnnotationForm';
+import {hideAnnotationForm} from "../../ui/uiSlice";
 
 interface MoveFormProps {
     readonly target: PythonDeclaration;
@@ -58,11 +59,11 @@ export const MoveForm: React.FC<MoveFormProps> = function ({ target }) {
                 ...data,
             }),
         );
-        dispatch(hideAnnotationForms());
+        dispatch(hideAnnotationForm());
     };
 
     const onCancel = () => {
-        dispatch(hideAnnotationForms());
+        dispatch(hideAnnotationForm());
     };
 
     // Rendering -------------------------------------------------------------------------------------------------------
