+++ api-editor/gui/src/features/annotations/forms/RenameForm.tsx
@@ -4,8 +4,9 @@ import { useForm } from 'react-hook-form';
 import { useAppDispatch, useAppSelector } from '../../../app/hooks';
 import { pythonIdentifierPattern } from '../../../common/validation';
 import PythonDeclaration from '../../packageData/model/PythonDeclaration';
-import { hideAnnotationForms, selectRenaming, upsertRenaming } from '../annotationSlice';
+import { selectRenaming, upsertRenaming } from '../annotationSlice';
 import { AnnotationForm } from './AnnotationForm';
+import {hideAnnotationForm} from "../../ui/uiSlice";
 
 interface RenameFormProps {
     readonly target: PythonDeclaration;
@@ -58,11 +59,11 @@ export const RenameForm: React.FC<RenameFormProps> = function ({ target }) {
                 ...data,
             }),
         );
-        dispatch(hideAnnotationForms());
+        dispatch(hideAnnotationForm());
     };
 
     const onCancel = () => {
-        dispatch(hideAnnotationForms());
+        dispatch(hideAnnotationForm());
     };
 
     // Rendering -------------------------------------------------------------------------------------------------------
