+++ api-editor/gui/src/features/packageData/selectionView/FunctionView.tsx
@@ -1,12 +1,12 @@
 import { Box, Heading, HStack, Stack, Text as ChakraText } from '@chakra-ui/react';
 import React from 'react';
 import { isEmptyList } from '../../../common/util/listOperations';
-import AnnotationDropdown from '../../annotations/AnnotationDropdown';
-import AnnotationView from '../../annotations/AnnotationView';
+import {AnnotationDropdown} from '../../annotations/AnnotationDropdown';
+import {AnnotationView} from '../../annotations/AnnotationView';
 import PythonFunction from '../model/PythonFunction';
 import PythonModule from '../model/PythonModule';
 import DocumentationText from './DocumentationText';
-import ParameterNode from './ParameterNode';
+import {ParameterNode} from './ParameterNode';
 import { useAppSelector } from '../../../app/hooks';
 import { selectCalledAfters } from '../../annotations/annotationSlice';
 
@@ -14,7 +14,7 @@ interface FunctionViewProps {
     pythonFunction: PythonFunction;
 }
 
-const FunctionView: React.FC<FunctionViewProps> = function ({ pythonFunction }) {
+export const FunctionView: React.FC<FunctionViewProps> = function ({ pythonFunction }) {
     const id = pythonFunction.pathAsString();
 
     // If more @calledAfter annotations can be added
@@ -73,5 +73,3 @@ const FunctionView: React.FC<FunctionViewProps> = function ({ pythonFunction }) 
         </Stack>
     );
 };
-
-export default FunctionView;
